refractor:add code refactor dashboard part 1
